inventario/requirements.txt
django>=5.0,<6.0
google-generativeai>=0.7.0
python-dotenv>=1.1.0

